@@ -10,15 +10,15 @@
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/cookie": "^0.5.1",
 		"sass": "^1.70.0",
-		"svelte": "^4.0.5",
-		"svelte-check": "^3.4.3",
-		"svelte-preprocess": "^5.0.4",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.0"
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"svelte-preprocess": "^6.0.0",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module",
 	"dependencies": {
